add cleanCoverage()
